requirements.txt
# HTTP requests
httpx==0.27.0

# Data analysis
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0

# Visualization (optional - for local analysis)
plotly==5.18.0

# Environment variables
python-dotenv==1.0.0

# Date handling
python-dateutil==2.8.2

# Image processing (for avatar caching)
Pillow==10.2.0

